adding script for project
This is a script that inputs local text files and outputs Patient IDs and UMLS terms.

This script splits text files within the working directory into lines, removes any lines with commonly 
negated terms and then extracts UMLS information from the remaining positive lines. This script outputs Patient IDs, 
given that the name of the file is the Patient ID, and UMLS information for non-negated text lines. The default 
is to also output each line that has positive terms. This can easily be commented out if desired. 
This script is meant for use with UCSF clinical notes. Please see the README file for information on the UMLS
metathesaurus and QuickUMLS installations.